Fixed a transparency issue with certain themes.
I use the MediterraneanDarkest theme with Mint. They worked very well
with Mint 16 Cinnamon. There was a regression with Mint 17 where the
grid container of the cinnamon-screensaver would be filled with a
background color.

This patch addresses this concern by overriding the background color
of the grid container with a transparent color.