Reorganize files, generate dbus proxies and add them to libcscreensaver so we can more easily access properties and signals of the various interfaces we're using, and be able to maintain a cleaner python layer (in dbusdepot).
A few remaining items:

Implement the screensaver service as a GDBus-based one, use GDBus
proxy in cinnamon-screensaver-command, and use one for the media-keys
handler.  This will let us eliminate any dependency on python3-dbus,
which seems a bit unreliable.  Using GLib/Gio stuff plays nicer with
the rest of the construct as well.