cs-notification-watcher.c: change what messages get passed on and what don't - due to misunderstanding of the documentation for this function.
The docs for GDBusMessageFilterFunction say to return the message
always, in order to just passively observe events (which is all we
want in this case - we're just counting notifications.)

What I mistook was that to return or not to return would affect
whether *other* dbus clients (programs) received the message in-turn.

This is not the case - these messages are for our use only.

Returning NULL when receiving a notification message tells dbus not
to propagate the message to any more filters in our process *only*.
It also lets dbus move on and send the message to other subscribers
(most importantly, Cinnamon, which is the usual target of libnotify
when it's running.)

It's important to pass on any non-relevant (to our interest in
notifications here) incoming message so that their proper handlers
can act on and respond to them (such as our session idle listener.)

This was preventing the idle signal from cinnamon-session from
ever being received, so the screensaver would never activate, unless
explicitly commanded to via cinnamon-screensaver-command.