Fix pam authentication to work with multi-factor, fingerprint, etc..
The new authentication in cinnamon-desktop-utils was far too simplistic,
it didn't allow for multiple pluggable modules to run (for example, password,
followed by a 2-factor authentication code).  We'll use the old screensaver
auth code instead - this ports gs-auth-pam.c and gs-auth.h from the old
screensaver, we add cinnamon-screensaver-pam-helper.c and authClient.py to
interface with it via pipes.

This lets us isolate our authentication code to a small executable instead
of the entire application.

Unfortunately it seemed better at this point to keep all of this stuff in
one package, so the authentication code will be removed in cinnamon-desktop.

Tested with:
- generic fingerprint reader (using fprintd and fingerprint-gui) - it works
  fine, but the ui overlay for fingerprint-gui doesn't currently appear over
  the unlock screen (though you may see it temporarily).
- google-authenticator - working as expected - prompt changes, allows code
  input
- duo unix - could not get to work, though I don't think it's our issue.  I get
  a permission error for /etc/duo/pam_duo.conf (though it appears to be correct)

Also, fixes no-password users so locking will not occur