manager/stage: Allow a second activation trigger to override a previous.
This prevents scenarios where the screen was in the middle of an idle fade
and a laptop lid is closed.  Previously, the 15 second fade would
prevail (and tend not to finish before a suspend,) and end up being forced
to finish after wakeup, and also possibly not locking, depending on settings.

With this, a subsequent command will always take priority.