service, manager: always complete the lock invocation, whether we succeed or not, and whether we were already active or not.
Previously, the service lock command would only return if the
screensaver was *not* already active.  With normal settings this
was fine.  But when setting 'delay before locking' in the screensaver
settings, there was the potential for cinnamon-screensaver-command -l
to never exit, as the lock completion was only being called after
screensaver activation.

This change limits delayed completion to only instances where the
screensaver was not already active.  If the screensaver is active
when receiving the 'lock' command, the service will complete the
invocation immediately.

See inline notes for more details