monitorView.py: Improve plugin behavior, logging, during screensaver activation/deactivation.
There can be large state changes during suspend/resume cycles, so we
need to check and re-check that our socket is still valid for attaching
a plugin window to both before we spawn it, and before we attempt to
add a resulting plug to the monitorView socket.  We use the recommended
gtk method for checking this, and kill or delay the plugin if necessary
to prevent problems.