power widget: Ensure widget state is correct when becoming Active, instead of only updating at the next power state change.
If the computer was already on battery when the screensaver became
active, the icon potentially would not show, or only an empty frame
would display.

Additionally, display the battery icon while asleep if the percentage
of any battery is below 20% and not plugged in.  Previously, the
battery icon would only be shown on the lockscreen, regardless of
charge level.