monitorView: don't use a transition for the view stack.
We've been relying on telemetry from the "transition-running"
property to let us know when to proceed along the wake sequence.
The problem is, animations can be flaky - there are situations where
the monitorView may not be mapped, and animations are skipped in
case.  There was really no need for a gradual transition here, it
was just for aesthetics transitioning from a plugin to the unlock
wallpaper view.  The jolting way GtkPlug/Socket works makes this
fairly pointless, and we need to be able to count on continuing
past the stack change in order to raise and make visible the unlock
dialog.

This had a net effect of occasionally having an invisible unlock
dialog when coming back from suspend, particularly when a plugin
was selected.

Potentially fixes #288