info panel: Update the battery status when raising the lock screen.
We don't get a reliable signal to watch for returning from sleep or
hibernate, other than SimulateUserActivity from our dbus service.
This is called by cinnamon-settings-daemon power manager when it
handles resume actions.

Make sure to update the battery icon when this occurs to respond
to any power changes that occurred while the computer was asleep.

Fixes #294