A couple small install fixups. (#302)
* fix broken pkg-config file

The dependency name was misspelled and thus could never have been used
since it was originally created.

* make: don't use intltool on desktop file, as there is nothing to do

There are no translations in tree, so this is unneeded activity. Just
configure the end result directly in configure, and immediately install
it.

* Remove unworking code to disable pam or switch to BSD compatibility.

I can verify the code doesn't work, since there is not in fact a
cs-auth-bsdauth.c file to utilize.

As for disabling pam entirely, the current check will force an error if
you choose the option to disable it, which makes the option pointless.
And if you were to pass the configure check (which seems to be valid on
BSD) you will still end up trying to include the "cs-auth-.c" sources,
and then things which are not #ifdef'ed will fail to find important
functions.