debian: don't try to run a testsuite that does not exist
According to my local debian experts, if I override this magic
incantation it will stop trying to run tests.

This is a good thing, because there have never been tests. There is,
however, a files/Makefile.in which doesn't contain tests (which is expected)
and doesn't define the many rules expected by:
https://www.gnu.org/software/automake/manual/html_node/Third_002dParty-Makefiles.html

As such, running make check does try to run the check target, but fails
because files/Makefile.in does not contain its own no-op check target.

The po/ directory also complains that POTFILES.in is inconsistent, which
is fine since we use ./makepot instead, so nothing here is actually
being maintained. Therefore it makes no sense to try running a testsuite
just so it can successfully say "no test code, therefore nothing to do".

The only remaining question I have is why does it ever try running in
the first place (but only some of the time), but this is not the hill I
wish to die on, so, moving on...