media player: Fix async setup of player controls.
When cinnamon-screensaver ran as a daemon, the media pleyer watcher
had plenty of time to set up and identify valid players before the
screensaver would actually be used.

Now that this is an on-demand service, the watcher won't be ready
before the stage spawns, when it would normally be queried over
whether or not to show the controls.

This sets things up so the audio controls and mpris connection
can be set up at any point, and the ui will update as needed.