stage.py: Delay initialization of certain components during startup.
Some components - power- and audio-related - use dbus interfaces to
receive updates, and get set up asynchronously. Since the screensaver
is now spawned on-demand, instead of always running in the background,
these need some time for their setup to complete.

Ideally they should be listening for their respective connections to
complete, and eventually they will, but this will require quite a
bit more code changes and add potential for other regressions, so
a fixed delay is ok for now.

This fixes media controls and album art not being displayed.