authentication: improve interaction between the main process and the pam helper.
Eliminate any query/response behavior when authentication is
completed or cancelled. Previously authClient would tell the pam
helper to exit, and wait for a response that it was doing so.

Now, the pam helper will exit under two conditions: successful
authentication will cause it to exit on its own (after sending
the success message), or cancellation by the main process will send
a SIGTERM to the helper, which will cause the helper to exit.

In either case, the main process will never wait for a response
or process end confirmation, making sure the UI is never hung
up on this. The auth client will perform any resource cleanup
asynchronously once the pam helper process actually exits.

Simplify the pam-helper's life cycle.