Use xdo to feed escape keycodes into the focused window prior to attempting mouse and keyboard grabs.
A frequent barrier to the screensaver starting is a modal window,
application popup menu or cinnamon popup being left open when
the user steps away from their computer. This will have the effect
of cancelling or closing these if they exist.

It sends two keystrokes in the event of a popup menu being raised
by a modal dialog (the first will cancel the popup, the second
the dialog).

note: cinnamon popups could be addressed differently (expo/overview
modes already are), but this wasn't possible with application
popups.

Ref: #93