cinnamon-screensaver-main.py: Fix application theme provider.
We relied on the return value from Gtk.CssProvider.load_from_data
method to determine whether or not to apply the fallback theme.

The return value itself was deprecated, but still works in Gtk3.
Unfortunately a change was recently made in pygobject that overrides
this method, and since the return value is removed in Gtk4, it
was removed from this override, so is always interpreted as False.

Use try/except instead, as that method can still raise an error
upon failure.

Fixes #446

Ref:
https://gitlab.gnome.org/GNOME/pygobject/-/merge_requests/231